Implemented PollValue
PollValue will enable categorizers to have a kind of poll to decide
a value between a pool of possible values.
This can be used i.e. to determine the ts_start or ts_end value, since
each stream (gps data, sensor data) has a first (or last) point with a
different timestamp, and only the value with the lesser (or greatest)
timestamp must be put in the database.